+++ src/components/Layout/LayoutApp/LayoutAppContext.tsx
@@ -1,9 +1,9 @@
 import { createContext } from "react";
 
-const LayoutContext = createContext<{
+const LayoutAppContext = createContext<{
   prevPathname: React.MutableRefObject<string>;
 }>({
   prevPathname: { current: "" },
 });
 
-export default LayoutContext;
+export default LayoutAppContext;
