+++ src/components/Layout/LayoutApp/common.ts
@@ -1,0 +1,1 @@
+export const fullLayoutPathnames = ["/new", "/story/[storyId]"];
