+++ src/components/Layout/LayoutApp/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LayoutApp";
