+++ src/components/Story/StoryFeed.tsx
@@ -96,7 +96,7 @@ const StoryFeed: React.FC<{ id: string }> = ({ id }) => {
 
   return (
     <>
-      <div className="w-full p-4 grid grid-cols-2 sm:grid-cols-4 lg:grid-cols-5 gap-4">
+      <div className="w-full p-4 grid grid-cols-2 sm:grid-cols-3 gap-4">
         {stories?.map((story, index) => (
           <StoryItem
             key={story.id}
