+++ src/containers/Story/StoryQueueable.tsx
@@ -157,7 +157,7 @@ const StoryQueueable: React.FC<{ story: Story }> = ({ story }) => {
     (prev, data) => data
   );
   return (
-    <div className="grid grid-cols-2 lg:grid-cols-3 gap-2">
+    <div className="grid grid-cols-2 sm:grid-cols-3 gap-2">
       <StoryQueueableAdder story={story} />
       {story.queueable.map((userId) => (
         <StoryQueueableUser key={userId} storyId={story.id} userId={userId} />
